Refactor Game as independent instance
Before this refactor, the Game was a Singleton unestable thing.
Probably this isn't the final version of the Game as an independent
instance, but is a beginning :)

The motivation to do this change is to allow create independent
instances of a Game in runtime, for exemple, to run a Game in the
editor. Morehover, this change will help to isolate better some code
dependencies as SDL2 and how to manage GL context.