Refactor Game to allow embedded its graphics
First step to embedded Game into the editor.
The idea is the Editor will run the game itself or the Editor use the
Game (or very similar object) to edit or preview scenes.
Probably this also will be useful to change the way that the Editor
render the preview of the in-game GUI inspector.