Create a time manager for game
This is to centralize al the related to the calculation of the frame
rate and time scale.
In the Future this class should allow to configure the frame rate and,
will do calculate the best physics rate based on game frame rate?
Also this class could be appli a time scale (bullet time? pause gmae?)