Remove unnecessary geGame interface
Remove the redundant code related to the game instance.
The idea is change all the interfaces to reduce the extra layer there is
between GameObjects and Components.

Also this change is made because I change my mind related with
the "public" API and "private" API. Currently there is an artificial
layer that only make the code more difficult to maintance. The idea is move
all code in a low level API and then I will create a scripting component
that it will expose a high level API to create simple components don't
require a complex implementation or low level access.