Fix copy test data files in Conan build
Conan copy all the source files into the build directory to avoid
poluting the source directory.
The error was that the CMake remove the same folder that try to copy and
this cause errors during build process.
Moreover, the step removed wasn't needed because the copy_directory
always copy all files.